+++ yodeploy/test_integration/samples/nginx-app/deploy/hooks.py
@@ -1,8 +1,7 @@
 import getpass
+import os
 
-import mock
-
-from os.path import join
+from mock import Mock
 
 from yodeploy.hooks.nginx import NginxHostedApp
 from yodeploy.test_integration.helpers import tests_dir
@@ -10,11 +9,14 @@ from yodeploy.test_integration.helpers import tests_dir
 
 class Hooks(NginxHostedApp):
     # these overrides are only needed for testing
-    nginx = mock.Mock()
-    logs_path = join(tests_dir, 'filesys/var/log/nginx')
-    server_blocks_path = join(tests_dir, 'filesys/etc/nginx/sites-enabled')
+    logs_path = os.path.join(tests_dir, 'filesys/var/log/nginx')
+    server_blocks_path = os.path.join(
+        tests_dir, 'filesys/etc/nginx/sites-enabled'
+    )
     user = getpass.getuser()
     group = None
+
+    _reload_nginx = Mock()
 
 
 hooks = Hooks