+++ yodeploy/test_integration/samples/uwsgi-app/deploy/hooks.py
@@ -1,8 +1,7 @@
 import getpass
+import os
 
-import mock
-
-from os.path import join
+from mock import Mock
 
 from yodeploy.hooks.uwsgi import UwsgiHostedApp
 from yodeploy.test_integration.helpers import tests_dir
@@ -10,11 +9,12 @@ from yodeploy.test_integration.helpers import tests_dir
 
 class Hooks(UwsgiHostedApp):
     # these overrides are only needed for testing
-    uwsgi = mock.Mock()
-    logs_path = join(tests_dir, 'filesys/var/log/uwsgi')
-    config_path = join(tests_dir, 'filesys/etc/uwsgi')
+    logs_path = os.path.join(tests_dir, 'filesys/var/log/uwsgi')
+    config_path = os.path.join(tests_dir, 'filesys/etc/uwsgi')
     user = getpass.getuser()
     group = None
+
+    _reload_uwsgi = Mock()
 
 
 hooks = Hooks