+++ yodeploy/test_integration/test_nginx_hooks.py
@@ -1,4 +1,4 @@
-from os.path import join
+import os
 
 from yodeploy.test_integration.helpers import rm, tests_dir
 from yodeploy.test_integration.helpers import build_sample, deploy_sample
@@ -12,9 +12,12 @@ class TestNginxAppDeployment(DeployTestCase):
         deploy_sample('nginx-app')
 
     def tearDown(self):
-        rm(join(tests_dir, 'filesys/var/log/nginx/nginx-app-access.log'))
-        rm(join(tests_dir, 'filesys/var/log/nginx/nginx-app-error.log'))
-        rm(join(tests_dir, 'filesys/etc/nginx/sites-enabled/nginx-app'))
+        rm(os.path.join(
+            tests_dir, 'filesys/var/log/nginx/nginx-app-access.log'))
+        rm(os.path.join(
+            tests_dir, 'filesys/var/log/nginx/nginx-app-error.log'))
+        rm(os.path.join(
+            tests_dir, 'filesys/etc/nginx/sites-enabled/nginx-app'))
 
     def test_places_server_block(self):
         self.assertDeployed('etc/nginx/sites-enabled/nginx-app')