+++ yodeploy/test_integration/test_uwsgi_hooks.py
@@ -1,4 +1,4 @@
-from os.path import join
+import os
 
 from yodeploy.test_integration.helpers import rm, tests_dir
 from yodeploy.test_integration.helpers import build_sample, deploy_sample
@@ -12,8 +12,8 @@ class TestUwsgiAppDeployment(DeployTestCase):
         deploy_sample('uwsgi-app')
 
     def tearDown(self):
-        rm(join(tests_dir, 'filesys/var/log/uwsgi/usgi-app.log'))
-        rm(join(tests_dir, 'filesys/etc/uwsgi/uwsgi-app.ini'))
+        rm(os.path.join(tests_dir, 'filesys/var/log/uwsgi/usgi-app.log'))
+        rm(os.path.join(tests_dir, 'filesys/etc/uwsgi/uwsgi-app.ini'))
 
     def test_places_config_ini(self):
         self.assertDeployed('etc/uwsgi/uwsgi-app.ini')