+++ CHANGELOG.md
@@ -4,9 +4,9 @@ Change Log
 0.9.0
 -----
 
-* Move out apache hooks from `DjangoApp` to new `ApacheHostedDjangoApp`;
-* Rename `DjangoMultiSiteApp` into `ApacheHostedDjangoMultiSiteApp`
-* Add `NginxHostedApp`, `UwsgiHostedApp` and `NginxHostedDjangoApp` hooks
+* Move out apache hooks from `DjangoApp` to new `ApacheHostedDjangoApp`.
+* Rename `DjangoMultiSiteApp` into `ApacheHostedDjangoMultiSiteApp`.
+* Add `NginxHostedApp`, `UwsgiHostedApp` and `NginxHostedDjangoApp` hooks.
 
 
 0.8.5